+++ src/imp/span_imp.js
@@ -1,82 +1,83 @@
 import * as coerce from './coerce.js';
 import * as constants from '../constants';
+import _each from '../_each';
+import opentracing from 'opentracing';
 import { crouton_thrift } from '../platform_abstraction_layer'; // eslint-disable-line camelcase
 
-export default class SpanImp {
+export default class SpanImp extends opentracing.Span {
 
     // ---------------------------------------------------------------------- //
-    // OpenTracing Implementation
+    // opentracing.Span SPI
     // ---------------------------------------------------------------------- //
 
-    tracer() {
-        return this._tracer;
+    _tracer() {
+        return this._tracerImp;
     }
 
-    context() {
-        return this._context;
+    _context() {
+        return this._ctx;
     }
 
-    setOperationName(name) {
-        this._operation = `${name}`;
+    _setOperationName(name) {
+        this._operationName = `${name}`;
     }
 
-    addTags(keyValuePairs) {
-        for (let key in keyValuePairs) {
-            // NB: Older versions of IE don't support `String.prototype.startsWith()`
-            if (key.substr(0, constants.JOIN_ID_PREFIX.length) === constants.JOIN_ID_PREFIX) {
-                this.setJoinID(key, keyValuePairs[key]);
-            } else {
-                this._tags[key] = keyValuePairs[key];
-            }
-        }
+    _addTags(keyValuePairs) {
+        let self = this;
+        _each(keyValuePairs, (value, key) => {
+            self._tags[key] = value;
+        });
     }
 
-    /**
-     * The set of OpenTracing fields is:
-     * - 'event'
-     * - 'timestamp'
-     * - 'payload'
-     */
-    // Log record specified by fields
-    log(fields) {
-        const argumentType = typeof fields;
+    _log(keyValuePairs, timestamp) {
+        let self = this;
+        const argumentType = typeof keyValuePairs;
         if (argumentType !== 'object') {
-            this._tracer._error('Span.log() expects an object as its argument');
+            self._tracerImp._error('Span.log() expects an object as its first argument');
             return;
         }
 
-        let rec = this._tracer.log()
-            .span(this.guid())
-            .level(constants.LOG_STRING_TO_LEVEL[fields.level] || constants.LOG_INFO);
-
-        //
-        // OpenTracing attributes
-        //
-        if (fields.event !== undefined) {
-            rec.name(fields.event);
+        let tsMicros = null;
+        if (timestamp) {
+            tsMicros = timestamp * 1000;
+        } else {
+            tsMicros = self._tracerImp._platform.nowMicros();
         }
-        if (fields.timestamp !== undefined) {
-            // The incoming 'timestamp' field is in milliseconds. The internal
-            // timestamp is in microseconds.
-            rec.timestamp(fields.timestamp * 1000);
-        }
-        if (fields.payload !== undefined) {
-            rec.payload(fields.payload);
-        }
-
-        //
-        // LightStep-specific attributes
-        //
-        if (fields.message !== undefined) {
-            rec.message(fields.message);
-        }
-        if (fields.timestamp_micros !== undefined) {
-            rec.timestamp(fields.timestamp_micros);
-        }
-        rec.end();
+        let fields = [];
+        _each(keyValuePairs, (value, key) => {
+            if (!key || !value) {
+                return;
+            }
+            let keyStr = coerce.toString(key);
+            let valStr = null;
+            if (value instanceof Object) {
+                valStr = JSON.stringify(value, null, "  ")
+            } else {
+                valStr = coerce.toString(value);
+            }
+            if (keyStr.length > self._tracerImp._options.log_field_key_hard_limit) {
+                self._tracerImp._counters['logs.keys.over_limit']++;
+                keyStr = keyStr.substr(0, self._tracerImp._options.log_field_key_hard_limit) + "...";
+            }
+            if (valStr.length > self._tracerImp._options.log_field_value_hard_limit) {
+                self._tracerImp._counters['logs.values.over_limit']++;
+                valStr = valStr.substr(0, self._tracerImp._options.log_field_value_hard_limit) + "...";
+            }
+            fields.push(new crouton_thrift.KeyValue({
+                Key   : keyStr,
+                Value : valStr,
+            }));
+        });
+        let record = new crouton_thrift.LogRecord({
+            timestamp_micros : tsMicros,
+            fields           : fields,
+        });
+        self._log_records = self._log_records || [];
+        self._log_records.push(record);
+        self._tracerImp.emit('log_added', record);
     }
 
-    finish(finishTime) {
+    _finish(finishTime) {
         return this.end(finishTime);
     }
 
@@ -84,19 +85,21 @@ export default class SpanImp {
     // Private methods
     // ---------------------------------------------------------------------- //
 
-    constructor(tracer, spanContext) {
+    constructor(tracer, name, spanContext) {
+        super();
+
         console.assert(typeof tracer === 'object', 'Invalid runtime');  // eslint-disable-line no-console
 
-        this._tracer = tracer;
-        this._context = spanContext;
+        this._tracerImp = tracer;
+        this._ctx = spanContext;
         this._ended  = false;
 
-        this._operation   = '';
-        this._tags        = {};
-        this._joinIDs     = {};
-        this._beginMicros = tracer._platform.nowMicros();
-        this._endMicros   = 0;
-        this._errorFlag   = false;
+        this._operationName = name;
+        this._tags          = {};
+        this._beginMicros   = tracer._platform.nowMicros();
+        this._endMicros     = 0;
+        this._errorFlag     = false;
+        this._log_records   = null;
     }
 
     // ---------------------------------------------------------------------- //
@@ -104,16 +107,16 @@ export default class SpanImp {
     // ---------------------------------------------------------------------- //
 
     getOperationName() {
-        return this._operation;
+        return this._operationName;
     }
 
     // Getter only. The GUID is immutable once set internally.
     guid() {
-        return this._context._guid;
+        return this._ctx._guid;
     }
 
     traceGUID() {
-        return this._context._traceGUID;
+        return this._ctx._traceGUID;
     }
 
     parentGUID() {
@@ -155,23 +158,17 @@ export default class SpanImp {
         if (this._beginMicros > 0 && this._endMicros > 0) {
             micros = Math.floor((this._beginMicros + this._endMicros) / 2);
         } else {
-            micros = this._tracer._platform.nowMicros();
+            micros = this._tracerImp._platform.nowMicros();
         }
 
         let urlPrefix = constants.LIGHTSTEP_APP_URL_PREFIX;
-        let accessToken = encodeURIComponent(this._tracer.options().access_token);
+        let accessToken = encodeURIComponent(this._tracerImp.options().access_token);
         let guid = encodeURIComponent(this.guid());
         return `${urlPrefix}/${accessToken}/trace?span_guid=${guid}&at_micros=${micros}`;
     }
 
     getTags() {
         return this._tags;
-    }
-
-    setJoinID(key, value) {
-        this._tags[key] = value;
-        this._joinIDs[key] = true;
-        return this;
     }
 
     /**
@@ -197,111 +194,32 @@ export default class SpanImp {
         // for retroactively created spans that might not be possible to create
         // in real-time).
         if (this._endMicros === 0) {
-            this._endMicros = this._tracer._platform.nowMicros();
+            this._endMicros = this._tracerImp._platform.nowMicros();
         }
-        this._tracer._addSpanRecord(this._toThrift());
-    }
-
-    // Info log record with an optional payload
-    info(msg, payload) {
-        this._tracer.log()
-            .span(this.guid())
-            .level(constants.LOG_INFO)
-            .message(msg)
-            .payload(payload)
-            .end();
-    }
-
-    warn(msg, payload) {
-        this._tracer.log()
-            .span(this.guid())
-            .level(constants.LOG_WARN)
-            .message(msg)
-            .payload(payload)
-            .end();
-    }
-
-    error(msg, payload) {
-        this._tracer.log()
-            .span(this.guid())
-            .level(constants.LOG_ERROR)
-            .message(msg)
-            .payload(payload)
-            .end();
-    }
-
-    // Special case to format exception information a little bit more nicely
-    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error
-    exception(msg, exception) {
-        if (exception.message === undefined || exception.stack === undefined) {
-            return this.error(msg, exception);
-        }
-
-        let stack = exception.stack.split('\n');
-        this._tracer.log()
-            .span(this.guid())
-            .level(constants.LOG_ERROR)
-            .message(msg)
-            .payload({
-                name        : exception.name,
-                message     : exception.message,
-                filename    : exception.filename,
-                line_number : exception.lineNumber,
-                stack       : stack,
-            })
-            .end();
-    }
-
-    fatal(msg, payload) {
-        this._tracer.log()
-            .span(this.guid())
-            .level(constants.LOG_FATAL)
-            .message(msg)
-            .payload(payload)
-            .end();
+        this._tracerImp._addSpanRecord(this._toThrift());
     }
 
     _toThrift() {
-        // TODO: the backend understands join IDs and attributes.  The outer API
-        // understands trace context attributes and tags.  The mapping is a
-        // little confusing at the moment...
-
-        let joinIDs = [];
         let attributes = [];
-        for (let key in this._tags) {
-            if (key in this._joinIDs) {
-                this._addTagAsJoinID(joinIDs, key);
-            } else {
-                attributes.push(new crouton_thrift.KeyValue({
-                    Key   : coerce.toString(key),
-                    Value : coerce.toString(this._tags[key]),
-                }));
-            }
-        }
+        _each(this._tags, (value, key) => {
+            attributes.push(new crouton_thrift.KeyValue({
+                Key   : coerce.toString(key),
+                Value : coerce.toString(value),
+            }));
+        });
 
         let record = new crouton_thrift.SpanRecord({
             span_guid       : this.guid(),
             trace_guid      : this.traceGUID(),
-            runtime_guid    : this._tracer.guid(),
-            span_name       : this._operation,
-            join_ids        : joinIDs,
+            runtime_guid    : this._tracerImp.guid(),
+            span_name       : this._operationName,
             oldest_micros   : this._beginMicros,
             youngest_micros : this._endMicros,
             attributes      : attributes,
             error_flag      : this._errorFlag,
+            log_records     : this._log_records,
         });
         return record;
     }
 
-    // Helper to reduce duplicated code
-    _addTagAsJoinID(joinIDs, key) {
-        let value = this._tags[key];
-        if (value === undefined) {
-            return;
-        }
-        joinIDs.push(new crouton_thrift.TraceJoinId({
-            TraceKey : coerce.toString(key),
-            Value    : coerce.toString(value),
-        }));
-    }
 }