+++ test/suites/common/tracer.js
@@ -81,14 +81,14 @@ describe("Tracer", function() {
             spanContext.setBaggageItem('creditcard', 'visa');
 
             var carrier = {};
-            Tracer.inject(spanContext, Tracer.FORMAT_TEXT_MAP, carrier);
-            expect(carrier['ot-tracer-traceid']).to.equal(spanContext.imp()._traceGUID);
-            expect(carrier['ot-tracer-spanid']).to.equal(spanContext.imp()._guid);
+            Tracer.inject(spanContext, opentracing.FORMAT_TEXT_MAP, carrier);
+            expect(carrier['ot-tracer-traceid']).to.equal(spanContext._traceGUID);
+            expect(carrier['ot-tracer-spanid']).to.equal(spanContext._guid);
             expect(carrier['ot-baggage-footwear']).to.equal('sandals');
             expect(carrier['ot-baggage-creditcard']).to.equal('visa');
 
-            var extractedContext = Tracer.extract(Tracer.FORMAT_TEXT_MAP, carrier);
-            expect(extractedContext.imp()._guid).to.equal(spanContext.imp()._guid);
+            var extractedContext = Tracer.extract(opentracing.FORMAT_TEXT_MAP, carrier);
+            expect(extractedContext._guid).to.equal(spanContext._guid);
             expect(extractedContext.getBaggageItem('footwear')).to.equal('sandals');
             expect(extractedContext.getBaggageItem('creditcard')).to.equal('visa');
         });
@@ -100,14 +100,14 @@ describe("Tracer", function() {
             spanContext.setBaggageItem('creditcard', 'visa');
 
             var carrier = {};
-            Tracer.inject(spanContext, Tracer.FORMAT_HTTP_HEADERS, carrier);
-            expect(carrier['ot-tracer-traceid']).to.equal(spanContext.imp()._traceGUID);
-            expect(carrier['ot-tracer-spanid']).to.equal(spanContext.imp()._guid);
+            Tracer.inject(spanContext, opentracing.FORMAT_HTTP_HEADERS, carrier);
+            expect(carrier['ot-tracer-traceid']).to.equal(spanContext._traceGUID);
+            expect(carrier['ot-tracer-spanid']).to.equal(spanContext._guid);
             expect(carrier['ot-baggage-footwear']).to.equal('sandals');
             expect(carrier['ot-baggage-creditcard']).to.equal('visa');
 
-            var extractedContext = Tracer.extract(Tracer.FORMAT_HTTP_HEADERS, carrier);
-            expect(extractedContext.imp()._guid).to.equal(spanContext.imp()._guid);
+            var extractedContext = Tracer.extract(opentracing.FORMAT_HTTP_HEADERS, carrier);
+            expect(extractedContext._guid).to.equal(spanContext._guid);
             expect(extractedContext.getBaggageItem('footwear')).to.equal('sandals');
             expect(extractedContext.getBaggageItem('creditcard')).to.equal('visa');
         });
@@ -122,7 +122,7 @@ describe("Tracer", function() {
 
         it("should extract from an empty text map carrier correctly" , function() {
             var carrier = {};
-            var extractedContext = Tracer.extract(Tracer.FORMAT_TEXT_MAP, carrier);
+            var extractedContext = Tracer.extract(opentracing.FORMAT_TEXT_MAP, carrier);
             expect(extractedContext).to.be.null;
         });
     });