+++ test/suites/node/reporting.js
@@ -2,21 +2,21 @@ var path = require("path");
 var fs   = require("fs");
 var child_process = require("child_process");
 var FileTransport = require("../../util/file_transport");
-var LightStep = require('../../..');
+var lightstep = require('../../..');
 
 describe("Reporting loop", function() {
     it("should flush at a regular interval", function(done) {
         this.timeout(5000);
 
         var transport = new FileTransport(path.join(__dirname, '../../results/report_flush_loop.json'));
-        var tracer = Tracer.initNewTracer(LightStep.tracer({
+        var tracer = new lightstep.Tracer({
             access_token                  : "{your_access_token}",
             component_name                : "api-javascript/unit-test/report_flush_loop",
             max_reporting_interval_millis : 10,
             override_transport            : transport,
             disable_reporting_loop        : false,
             delay_initial_report_millis   : 0,
-        }));
+        });
 
         var count = 0;
         var timer = setInterval(function() {
@@ -30,7 +30,7 @@ describe("Reporting loop", function() {
 
                 // Conservatively check below the theoretical values since this
                 // test inherently has timing issues.
-                expect(reqs.logRecordCount()).to.be.at.least(10);
+                expect(reqs.spanRecordCount()).to.be.at.least(10);
                 expect(reqs.reportCount()).to.be.at.least(2);
                 return done();
             }
@@ -47,17 +47,17 @@ describe("Final report", function () {
         child.on('close', function() {
             var reqs = util.requestsFromFile(reportFile);
 
-            expect(reqs.logRecordCount()).to.be.at.least(10);
-            expect(reqs.hasLogMessage("log0")).to.be.true;
-            expect(reqs.hasLogMessage("log1")).to.be.true;
-            expect(reqs.hasLogMessage("log2")).to.be.true;
-            expect(reqs.hasLogMessage("log3")).to.be.true;
-            expect(reqs.hasLogMessage("log4")).to.be.true;
-            expect(reqs.hasLogMessage("log5")).to.be.true;
-            expect(reqs.hasLogMessage("log6")).to.be.true;
-            expect(reqs.hasLogMessage("log7")).to.be.true;
-            expect(reqs.hasLogMessage("log8")).to.be.true;
-            expect(reqs.hasLogMessage("log9")).to.be.true;
+            expect(reqs.spanRecordCount()).to.be.at.least(10);
+            expect(reqs.hasSpanRecord("test_span_0")).to.be.true;
+            expect(reqs.hasSpanRecord("test_span_1")).to.be.true;
+            expect(reqs.hasSpanRecord("test_span_2")).to.be.true;
+            expect(reqs.hasSpanRecord("test_span_3")).to.be.true;
+            expect(reqs.hasSpanRecord("test_span_4")).to.be.true;
+            expect(reqs.hasSpanRecord("test_span_5")).to.be.true;
+            expect(reqs.hasSpanRecord("test_span_6")).to.be.true;
+            expect(reqs.hasSpanRecord("test_span_7")).to.be.true;
+            expect(reqs.hasSpanRecord("test_span_8")).to.be.true;
+            expect(reqs.hasSpanRecord("test_span_9")).to.be.true;
 
             expect(reqs.spanRecordCount()).to.be.at.least(1);